+++ classic/5.0/src/AbsBeamline/Collimator.h
@@ -69,9 +69,9 @@ public:
     /// Return the vertical half-aperture.
     virtual double getYsize() const {return b_m;}
 
-    virtual bool apply(const int &i, const double &t, double E[], double B[]);
+    virtual bool apply(const size_t &i, const double &t, double E[], double B[]);
 
-    virtual bool apply(const int &i, const double &t, Vector_t &E, Vector_t &B);
+    virtual bool apply(const size_t &i, const double &t, Vector_t &E, Vector_t &B);
 
     virtual bool apply(const Vector_t &R, const Vector_t &centroid, const double &t, Vector_t &E, Vector_t &B);
