+++ classic/5.0/src/AbsBeamline/Diagnostic.cpp
@@ -48,11 +48,11 @@ void Diagnostic::accept(BeamlineVisitor &visitor) const {
     visitor.visitDiagnostic(*this);
 }
 
-bool Diagnostic::apply(const int &i, const double &t, double E[], double B[]) {
+bool Diagnostic::apply(const size_t &i, const double &t, double E[], double B[]) {
     return false;
 }
 
-bool Diagnostic::apply(const int &i, const double &t, Vector_t &E, Vector_t &B) {
+bool Diagnostic::apply(const size_t &i, const double &t, Vector_t &E, Vector_t &B) {
     return false;
 }
