+++ classic/5.0/src/AbsBeamline/Multipole.cpp
@@ -144,7 +144,7 @@ void Multipole::addKT(int i, double t, Vector_t &K) {
     K += Vector_t(cf * dx, -cf * dy, 0.0);
 }
 
-bool Multipole::apply(const int &i, const double &t, double E[], double B[]) {
+bool Multipole::apply(const size_t &i, const double &t, double E[], double B[]) {
     Vector_t Ev(0, 0, 0), Bv(0, 0, 0);
     Vector_t Rt(RefPartBunch_m->getX(i), RefPartBunch_m->getY(i), RefPartBunch_m->getZ(i));
     if(apply(Rt, Vector_t(0.0), t, Ev, Bv)) return true;
@@ -159,7 +159,7 @@ bool Multipole::apply(const int &i, const double &t, double E[], double B[]) {
     return false;
 }
 
-bool Multipole::apply(const int &i, const double &t, Vector_t &E, Vector_t &B) {
+bool Multipole::apply(const size_t &i, const double &t, Vector_t &E, Vector_t &B) {
     Vector_t temp(RefPartBunch_m->getX(i), RefPartBunch_m->getY(i), RefPartBunch_m->getZ(i));
 
     const Vector_t &R(temp);
